Related work items: #4381

+++ PrimeApps.Studio/Controllers/TemplateController.cs
@@ -70,7 +70,7 @@ namespace PrimeApps.Studio.Controllers
         public async Task<IActionResult> Get(int id)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.View))
-                return StatusCode(403);
+                return Forbid();
 
             var template = await _templateRepostory.GetById(id);
 
@@ -85,7 +85,7 @@ namespace PrimeApps.Studio.Controllers
             GetAll([FromUri] TemplateType templateType) //JObject obj)// = TemplateType.NotSet, [FromQuery(Name = "moduleName")]string moduleName = "")	
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.View))
-                return StatusCode(403);
+                return Forbid();
 
             var templates = await _templateRepostory.GetAll(templateType, LanguageType.NotSet, false); //, moduleName);
 
@@ -97,7 +97,7 @@ namespace PrimeApps.Studio.Controllers
             TemplateType excelType = TemplateType.NotSet, string moduleName = "")
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.View))
-                return StatusCode(403);
+                return Forbid();
 
             var templates = await _templateRepostory.GetAllList(LanguageType.NotSet, type, excelType, moduleName);
 
@@ -108,7 +108,7 @@ namespace PrimeApps.Studio.Controllers
         public async Task<IActionResult> Create([FromBody] TemplateBindingModel template)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.Create))
-                return StatusCode(403);
+                return Forbid();
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
@@ -132,7 +132,7 @@ namespace PrimeApps.Studio.Controllers
         public async Task<IActionResult> CreateExcel([FromBody] TemplateBindingModel template)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.Create))
-                return StatusCode(403);
+                return Forbid();
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
@@ -155,7 +155,7 @@ namespace PrimeApps.Studio.Controllers
         public async Task<IActionResult> Update(int id, [FromBody] TemplateBindingModel template)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.Update))
-                return StatusCode(403);
+                return Forbid();
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
@@ -179,7 +179,7 @@ namespace PrimeApps.Studio.Controllers
         public async Task<IActionResult> Delete(int id, [FromQuery]bool isAppTemplate = false)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.Delete))
-                return StatusCode(403);
+                return Forbid();
 
             if (isAppTemplate)
             {
@@ -187,12 +187,12 @@ namespace PrimeApps.Studio.Controllers
 
                 if (templateEntity == null)
                     return NotFound();
-                
+
                 if (templateEntity.SystemCode == "email_confirm" || templateEntity.SystemCode == "password_reset")
                     return Forbid();
-                
+
                 await _appDraftTemplateRepository.DeleteSoft(templateEntity);
-                
+
                 return Ok();
             }
             else
@@ -202,18 +202,21 @@ namespace PrimeApps.Studio.Controllers
                 if (templateEntity == null)
                     return NotFound();
 
+                if (templateEntity.SystemType == SystemType.System)
+                    return Forbid();
+
                 await _templateRepostory.DeleteSoft(templateEntity);
 
                 return Ok();
             }
-          
+
         }
 
         [Route("count"), HttpGet]
         public IActionResult Count([FromUri] TemplateType templateType)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.View))
-                return StatusCode(403);
+                return Forbid();
 
             var count = _templateRepostory.Count(templateType);
             return Ok(count);
@@ -223,7 +226,7 @@ namespace PrimeApps.Studio.Controllers
         public IActionResult Find(ODataQueryOptions<Template> queryOptions, [FromUri]TemplateType templateType = 0)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.View))
-                return StatusCode(403);
+                return Forbid();
 
             var temps = _templateRepostory.Find(templateType);
 
@@ -235,7 +238,7 @@ namespace PrimeApps.Studio.Controllers
         public async Task<IActionResult> CreateAppEmailTemplate([FromBody] AppTemplateBindingModel template)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.Create))
-                return StatusCode(403);
+                return Forbid();
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
@@ -261,26 +264,26 @@ namespace PrimeApps.Studio.Controllers
         public async Task<IActionResult> UpdateAppEmailTemplate(int id, [FromBody] AppTemplateBindingModel template)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.Update))
-                return StatusCode(403);
+                return Forbid();
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-   
-                var templateEntity = await _appDraftTemplateRepository.Get(id);
 
-                if (templateEntity == null)
-                    return NotFound();
+            var templateEntity = await _appDraftTemplateRepository.Get(id);
 
-                TemplateHelper.UpdateEntity(null, null, null, template, templateEntity, true);
-                await _appDraftTemplateRepository.Update(templateEntity);
-                return Ok(templateEntity);          
+            if (templateEntity == null)
+                return NotFound();
+
+            TemplateHelper.UpdateEntity(null, null, null, template, templateEntity, true);
+            await _appDraftTemplateRepository.Update(templateEntity);
+            return Ok(templateEntity);
         }
 
         [Route("count_app_email_template"), HttpGet]
         public async Task<IActionResult> CountAppTemplate([FromUri] string currentAppName)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.View))
-                return StatusCode(403);
+                return Forbid();
 
             var app = await _appDraftRepository.GetByName(currentAppName.ToLower());
             var count = app != null ? _appDraftTemplateRepository.Count(app.Id) : 0;
@@ -289,25 +292,23 @@ namespace PrimeApps.Studio.Controllers
         }
 
         [Route("find_app_email_template")]
-        public IActionResult FindAppEmailTemplate(ODataQueryOptions<AppDraftTemplate> queryOptions)
+        public IActionResult FindAppEmailTemplate(ODataQueryOptions<AppDraftTemplate> queryOptions, [FromUri]int appId)
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.View))
-                return StatusCode(403);
+                return Forbid();
 
-            var views = _appDraftTemplateRepository.Find();
+            var views = _appDraftTemplateRepository.Find(appId);
             var queryResults = (IQueryable<AppDraftTemplate>)queryOptions.ApplyTo(views, new ODataQuerySettings() { EnsureStableOrdering = false });
-            return Ok(new PageResult<AppDraftTemplate>(queryResults, Request.ODataFeature().NextLink,
-                Request.ODataFeature().TotalCount));
+            return Ok(new PageResult<AppDraftTemplate>(queryResults, Request.ODataFeature().NextLink, Request.ODataFeature().TotalCount));
         }
-
 
         [Route("get_all_by_app_id"), HttpGet]
         public async Task<IActionResult> GetAllByAppId()
         {
             if (!_permissionHelper.CheckUserProfile(UserProfile, "template", RequestTypeEnum.View))
-                return StatusCode(403);
+                return Forbid();
 
-            var templates = await _appDraftTemplateRepository.GetAll((int)AppId);
+            var templates = await _appDraftTemplateRepository.GetAllById((int)AppId);
 
             return Ok(templates);
         }
